--- a/hw10_program_optimization/old
+++ b/hw10_program_optimization/old
@@ -1,0 +1,11 @@
+goos: linux
+goarch: amd64
+pkg: github.com/a-klimenko/go-otus-hw/hw10_program_optimization
+cpu: Intel(R) Core(TM) i7-9750H CPU @ 2.60GHz
+BenchmarkGetDomainStat-12    	       2	 634587736 ns/op	308473480 B/op	 3145444 allocs/op
+BenchmarkGetDomainStat-12    	       2	 631814186 ns/op	308397756 B/op	 3145430 allocs/op
+BenchmarkGetDomainStat-12    	       2	 641418458 ns/op	308247692 B/op	 3145421 allocs/op
+BenchmarkGetDomainStat-12    	       2	 642778201 ns/op	308435632 B/op	 3145438 allocs/op
+BenchmarkGetDomainStat-12    	       2	 607667098 ns/op	308396240 B/op	 3145435 allocs/op
+PASS
+ok  	github.com/a-klimenko/go-otus-hw/hw10_program_optimization	12.913s